Integration for player

See merge request compsci308_2019spring/voogasalad_nomergeconflicts!87

+++ src/Configurations/src/Configs/GamePackage/Game.java
@@ -52,7 +52,7 @@ public class Game implements Updatable, EventHandlable, Configurable {
     }
 
     @Override
-    public void update(long ms) {
+    public void update(double ms) {
         myActiveLevel.update(ms);
         if(myActiveLevel.noMoreEnemiesLeft()) {
             currentLevelOver = true;
